+++ src/pcasph.jl
@@ -93,5 +93,6 @@ function pcasph!(X::Matrix, weights::Weight;
     T .= T[:, s]
     P .= P[:, s]
     sv .= sv[s]
-    Pca(T, P, sv, xmeans, xscales, weights, nothing, kwargs, par) 
+    Pca(T, P, sv, xmeans, xscales, weights, 
+        nothing, kwargs, par) 
 end